feat: consume a reader rather
As written, data is fed to the parser using `feed_str`, which requires
ALL of the data to be fed before you can start listening for events.

This will consume a lot of memory, and is counter to the whole streaming
philosophy if you have say, a 10GB XML file that you want to parse.

This commit tweaks the API slightly to accept any `Read`er, and you can
start listening to events immediately.